Makefile: Correctly create IPC directories
The command that is supposed to create the directory structure for
IPC messaging is written using brace expansion, despite not being
a shell script. Because GNU Make does not support brace expansion,
it ends up creating an incorrect directory structure.

This commit changes the install rule of the Makefile to avoid brace
expansion.