utils: Allow ipc-sshtunnel and friends to be used for uipc messages
The scripts for tunneling IPC messages over SSH cannot be used with
uipc messages, even though the two modules share the same interface.

This commit modifies the scripts ipc-sshtunnel, ipc-inject, and
ipc-tap, so that they can handle uipc as well as ipc messages by
adding a command line option for specifying the ipc protocol to be
tunneled.