utils: Make ipc-tap and ipc-inject use the same endpoint
When ipc-sshtunnel is used to tunnel a pubsub topic in both directions,
ipc-tap will receive messages sent from the ipc-inject instance of the
same tunnel. This makes it impossible to use ipc-sshtunnel to relay a
topic in both directions.

This commit modifies ipc-tap and ipc-inject so that the caller may pass
an endpoint name that the scripts will use for IPC communication. This
further modifies ipc-sshtunnel so that it passes the same endpoint name
to ipc-tap and ipc-inject, causing the same endpoint to be used for
sending and receiving of messages. This causes messages from ipc-inject
not to be sent to ipc-tap.