include/ipc: Correctly assign `topics' in ipc_endpoint_subscribe()
The `topics' argument of ipc_endpoint_subscribe() is not assigned
from an array, causing the function to subscribe only to the first
topic that is passed to it.

This commit fixes the assignment of the `topics' argument of
ipc_endpoint_subscribe() so that the function will subscribe an
endpoint to more than one topic.