debian: Add dependencies for JSON Schema validation script
The script for JSON Schema validation requires the python3 and
python3-jsonschema packages to be installed, but because the
toolbox debian package does not depend on them, they may not be
installed.

This commit makes the toolbox debian package depend on python3
and python3-jsonschema, making sure that the JSON Schema validator
can be executed.