Makefile: Install JSON Schema files and validator
When installing toolbox, neither the JSON Schemas for IPC messages
nor the validator are installed. Thus, it is not possible to use the
validator to validate received messages. However, the validator is
necessary to implement incoming message validation in the ipc module.

This commit modifies the Makefile so that the JSON Schema files and
validator are installed by the `install' target.